[wip] combine testing
